feat: update Vite configuration for backend proxy, added CORS config, and API documentation
