+++ client/vite.config.ts
@@ -1,15 +1,15 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 
-const port = Number(process.env.VITE_PORT) || 8080;
-const backendOrigin = process.env.VITE_BACKEND_ORIGIN_DEV || "http://localhost:8080/api";
+const backendOrigin = process.env.VITE_BACKEND_ORIGIN;
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
   server: {
-    port,
     proxy: {
+      // Redirect all requests to the backend server
+      // example: localhost:5173/api/login -> localhost:8080/api/login
       "/api": {
         target: backendOrigin,
         changeOrigin: true,
@@ -19,7 +19,6 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": "/src",
-      "@shared": "../shared",
     },
   },
 });
