+++ shared/types/index.ts
@@ -54,15 +54,3 @@ export type NoteDTO = {
   createdAt?: Date;
   updatedAt?: Date;
 };
-
-export const httpCodes = {
-  SUCCESS: 200,
-  CREATED: 201,
-  ACCEPTED: 202,
-  NO_CONTENT: 204,
-  NOT_FOUND: 404,
-  BAD_REQUEST: 400,
-  UNAUTHORIZED: 401,
-  FORBIDDEN: 403,
-  INTERNAL_SERVER_ERROR: 500,
-};
